Fix compare against unsigned int warnings

+++ src/int_map.cpp
@@ -6,8 +6,8 @@ IntMap::IntMap(IntMappable* raw) : Mat2<unsigned int>(raw->height(), raw->width(
 {
 	unsigned int wdth = raw->width();
 
-	for (int j = 0; j < raw->height(); j++) {
-		for (int i = 0; i < wdth; i++) {
+	for (size_t j = 0; j < raw->height(); j++) {
+		for (size_t i = 0; i < wdth; i++) {
 			setElem((j * wdth) +  i, raw->pixelInt(j, i));
 		}
 	}
@@ -19,8 +19,8 @@ IntMap::IntMap(const std::unique_ptr<IntMappable>& raw) : Mat2<unsigned int>(raw
 
 	auto fe = begin();
 	
-	for (int j = 0; j < raw->height(); j++) {
-		for (int i = 0; i < wdth; i++) {
+	for (size_t j = 0; j < raw->height(); j++) {
+		for (size_t i = 0; i < wdth; i++) {
 			// setElem(j*wdth + i, raw->pixelInt(j, i));
 			*(fe++) =  raw->pixelInt(j, i);
 		}
